OV-311: + extract video name input into component
