+++ frontend/src/bundles/studio/components/components.ts
@@ -3,3 +3,4 @@ export { PlayerControls } from './player-controls/player-controls.js';
 export { Timeline } from './timeline/timeline.js';
 export { VideoComponent } from './video/video.js';
 export { VideoMenu } from './video-menu/video-menu.js';
+export { VideoNameInput } from './video-name-input/video-name-input.js';
